Table changes and corrections

+++ src/Components/Charts/LineChart.jsx
@@ -19,7 +19,7 @@ const LineChartDetails = ({ datas }) => {
   // Transform tableData to the desired format
   const formattedData = datas?.tableData?.map((item) => ({
     name: item?.date?.split(" ")[0], // Extract date part only
-    data1: parseInt(
+    data1: parseFloat(
       item?.["blood_ketone_value"] ||
         item?.["hemoglobinValue"] ||
         item?.["bmi"] ||
